move ingress_link functions to their own file
